+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.0.0](https://github.com/daggy1234/dagpi.js/compare/v1.1.5...v2.0.0) (2021-09-18)
+
 ### [1.1.5](https://github.com/daggy1234/dagpi.js/compare/v1.1.4...v1.1.5) (2021-06-05)
 
 - cube